+++ src/ergotest/suite/runnable_function.ts
@@ -1,7 +1,7 @@
 // Copyright Titanium I.T. LLC. License granted under terms of "The MIT License."
 import { Milliseconds, RecursiveRunOptions } from "./test.js";
 import { RunResult } from "../results/test_result.js";
-import { ItFn, ItOptions } from "./test_suite.js";
+import { ItFn, ItOptions } from "../test_api.js";
 
 export type BeforeAfterDefinition = { name?: string[], options: ItOptions, fnAsync: ItFn };